requirements.txt
requests==2.24.0
lxml==4.6.1
Flask==1.1.2
beautifulsoup4==4.9.3
google==3.0.0
python-magic==0.4.22